Add mimalloc with explicit allocator strategy
Switches the project to use mimalloc as the default allocator across
platforms, while keeping allocator ownership explicit and preserving
malloc-based tooling (Valgrind/ASan) via a build-time switch. This also
moves our tracking allocation functions xmalloc, xfree etc. to a
dedicated memory.c file to keep main.c free of memory specific concerns
like this.

Rationale

- We have observed hitching on Windows and Linux (more pronounced on
Windows); in testing, using mimalloc reduced hitching and improved
overall performance relative to the default system allocators.
- Mixing allocators is acceptable only when ownership is clear. This
change prioritizes explicit boundaries to avoid cross-allocator
free/realloc bugs.

Allocator strategy

- **Our code:** `MALLOC` / `CALLOC` / `REALLOC` / `FREE` / `STRDUP`
macros map to mimalloc when enabled, keeping allocator choice explicit
at call sites.
- **Third-party libraries (explicit configuration where supported):**
  - **SDL:** uses mimalloc via `SDL_SetMemoryFunctions()`.
  - **libpng:** uses custom allocators via `png_create_read_struct_2()`
  and `png_create_write_struct_2()`.
  - **libzip:** remains on the system allocator (controlled exception)
  to avoid implicit allocator substitution across shared-library
  boundaries.

Build / tooling

- Add `USE_MIMALLOC` build flag:
  - Default: enabled
  - Disabled for Valgrind/ASan builds to ensure malloc-based tooling
  works as expected.
- Update build scripts and CI environments to provide mimalloc where
required.

Testing

- Verified startup/shutdown and core flows on Windows, macOS, and Linux.
- Observed reduced hitching on Windows and Linux under previously
problematic workloads.
