+++ .github/workflows/build.yml
@@ -37,7 +37,8 @@ jobs:
             libsdl2-mixer-dev \
             libpng-dev \
             libzip-dev \
-            zlib1g-dev
+            zlib1g-dev \
+            libmimalloc-dev
 
       - name: Setup Rust
         uses: actions-rust-lang/setup-rust-toolchain@v1
@@ -91,6 +92,7 @@ jobs:
             mingw-w64-clang-x86_64-SDL2_mixer
             mingw-w64-clang-x86_64-libzip
             mingw-w64-clang-x86_64-dwarfstack
+            mingw-w64-clang-x86_64-mimalloc
             make
             zip
 
@@ -126,7 +128,7 @@ jobs:
 
       - name: Install dependencies
         run: |
-          brew install sdl2 sdl2_mixer libzip zlib imagemagick
+          brew install sdl2 sdl2_mixer libzip zlib mimalloc imagemagick
 
       - name: Setup Rust
         uses: actions-rust-lang/setup-rust-toolchain@v1
