+++ .github/workflows/quality.yml
@@ -56,7 +56,8 @@ jobs:
           libsdl2-mixer-dev \
           libpng-dev \
           libzip-dev \
-          zlib1g-dev
+          zlib1g-dev \
+          libmimalloc-dev
 
     - name: Setup Rust
       uses: actions-rust-lang/setup-rust-toolchain@v1
@@ -155,7 +156,8 @@ jobs:
           libsdl2-mixer-dev \
           libpng-dev \
           libzip-dev \
-          zlib1g-dev
+          zlib1g-dev \
+          libmimalloc-dev
 
     - name: Setup Rust
       uses: actions-rust-lang/setup-rust-toolchain@v1
